improve: refactoring
