+++ src/app/(home)/components/Fashion/Fashion.tsx
@@ -2,7 +2,7 @@ import ArrowUpRightIcon from "@/shared/components/icons/Arrows/ArrowUpRightIcon"
 import { Button } from "@/shared/components/UI/Button";
 import Container from "@/shared/components/UI/Container";
 import Section from "@/shared/components/UI/Section";
-import FashionCard from "@/features/Fashion/components/FashionCard";
+import FashionCard from "@/features/Products/components/Cards/FashionCard";
 
 const Fashion = () => {
   return (
