+++ src/features/Products/components/Cards/FashionCard.tsx
