+++ src/features/Products/components/Cards/ProductServerCard.tsx
@@ -13,6 +13,9 @@ import { useLocalStorage } from "@/shared/hooks/useLocalStorage";
 import { useQueryClient } from "@tanstack/react-query";
 import { getProductByIdQueryOptions } from "../../api/productsApi";
 
+import { Skeleton } from "@/shared/components/UI/Skeleton";
+import { cn } from "@/shared/utils/cn";
+
 const ProductServerCard = ({ product }: { product: ProductServer }) => {
   const queryClient = useQueryClient();
   const [_, setRecentProducts] = useLocalStorage<number[]>(
@@ -103,3 +106,13 @@ const ProductServerCard = ({ product }: { product: ProductServer }) => {
 };
 
 export default ProductServerCard;
+
+export const ProductCardLoader = ({ className }: { className?: string }) => {
+  return (
+    <div className={cn("flex flex-col gap-4 w-full h-full", className)}>
+      <Skeleton className="w-full h-full" />
+      <Skeleton className="w-3/5 h-[18px]" />
+      <Skeleton className="w-2/5 h-[18px]" />
+    </div>
+  );
+};
